pulled out versioning into its own API, corrected unit tests, made project 100% PEP8 compliant
