--- config.py
@@ -1,9 +1,0 @@
-config = {
-    'sqlalchemy' : {
-        'url'           : 'sqlite:////tmp/meniscus.db',
-        'echo'          : False,
-        'echo_pool'     : False,
-        'pool_recycle'  : 3600,
-        'encoding'      : 'utf-8'
-    }
-}
