+++ meniscus/personas/worker/correlation/app.py
@@ -3,7 +3,7 @@ import falcon
 from meniscus.personas.worker.register_online import RegisterWorkerOnline
 from meniscus.proxy import NativeProxy
 from meniscus.api.correlation.resources import PublishMessageResource
-from meniscus.api.correlation.resources import VersionResource
+from meniscus.api.version.resources import VersionResource
 
 
 def start_up(cfg=dict()):
