+++ meniscus/tests/api/pairing/resources_test.py
@@ -5,35 +5,13 @@ from mock import MagicMock
 from mock import patch
 import falcon
 
-from meniscus.api.pairing.resources \
-    import VersionResource, PairingConfigurationResource
+from meniscus.api.pairing.resources import PairingConfigurationResource
 
 
 def suite():
     suite = unittest.TestSuite()
-    suite.addTest(WhenTestingPairingVersionResource())
     suite.addTest(WhenTestingPairingConfigurationResource())
     return suite
-
-
-class WhenTestingPairingVersionResource(unittest.TestCase):
-
-    def setUp(self):
-        self.req = MagicMock()
-        self.resp = MagicMock()
-        self.resource = VersionResource()
-
-    def test_should_return_200_on_get(self):
-        self.resource.on_get(self.req, self.resp)
-        self.assertEqual(falcon.HTTP_200, self.resp.status)
-
-    def test_should_return_version_json(self):
-        self.resource.on_get(self.req, self.resp)
-
-        parsed_body = json.loads(self.resp.body)
-
-        self.assertTrue('v1' in parsed_body)
-        self.assertEqual('current', parsed_body['v1'])
 
 
 class WhenTestingPairingConfigurationResource(unittest.TestCase):
