+++ meniscus/tests/api/tenant/resources_test.py
@@ -13,7 +13,6 @@ from meniscus.api.tenant.resources import HostProfilesResource
 from meniscus.api.tenant.resources import TenantResource
 from meniscus.api.tenant.resources import TokenResource
 from meniscus.api.tenant.resources import UserResource
-from meniscus.api.tenant.resources import VersionResource
 from meniscus.data.model.tenant import EventProducer
 from meniscus.data.model.tenant import Host
 from meniscus.data.model.tenant import HostProfile
@@ -25,7 +24,6 @@ from meniscus.openstack.common import jsonutils
 def suite():
 
     test_suite = unittest.TestSuite()
-    test_suite.addTest(WhenTestingVersionResource())
 
     test_suite.addTest(WhenTestingTenantResourceOnPost())
 
@@ -110,26 +108,6 @@ class TestingTenantApiBase(unittest.TestCase):
 
     def _set_resource(self):
         pass
-
-
-class WhenTestingVersionResource(unittest.TestCase):
-
-    def setUp(self):
-        self.req = MagicMock()
-        self.resp = MagicMock()
-        self.resource = VersionResource()
-
-    def test_should_return_200_on_get(self):
-        self.resource.on_get(self.req, self.resp)
-        self.assertEqual(falcon.HTTP_200, self.resp.status)
-
-    def test_should_return_version_json(self):
-        self.resource.on_get(self.req, self.resp)
-
-        parsed_body = jsonutils.loads(self.resp.body)
-
-        self.assertTrue('v1' in parsed_body)
-        self.assertEqual('current', parsed_body['v1'])
 
 
 class WhenTestingTenantResourceOnPost(TestingTenantApiBase):
@@ -449,12 +427,20 @@ class WhenTestingEventProducersResourceValidate(TestingTenantApiBase):
         self.resource = EventProducersResource(self.db_handler)
 
     def test_should_throw_exception_for_bad_durable_val(self):
-        body = {'name': 'ep_name', 'pattern': 'syslog', 'durable': "bad_data"}
+        body = {
+            'name': 'ep_name',
+            'pattern': 'syslog',
+            'durable': "bad_data"
+        }
         with self.assertRaises(falcon.HTTPError):
             self.resource._validate_req_body_on_post(body)
 
     def test_should_throw_exception_for_bad_encrypted_val(self):
-        body = {'name': 'ep_name', 'pattern': 'syslog', 'encrypted': "bad_data"}
+        body = {
+            'name': 'ep_name',
+            'pattern': 'syslog',
+            'encrypted': "bad_data"
+        }
         with self.assertRaises(falcon.HTTPError):
             self.resource._validate_req_body_on_post(body)
